Add URDF
Adds a basic robot description including a URDF, RViz config, and a launch file to publish the URDF.

The URDF defines the robot's base link and links for the DVL, USBL, FLS, and Voyis stereo cameras.

The RViz config provides a default visualization setup.

The launch file publishes the URDF to the /robot_description topic using the robot_state_publisher.